Fix foldable mispredict
